// whirlpool/whirlpoolemea/worldpay-connector/node/typings/configs.d.ts
export interface configs {
  vtexAppKey: string,
  vtexAppToken: string,
  wpUrl: string,
  merchantCode: string,
  installationID: string,
  wpusername: string,
  wppassword: string,
  paymentAppName: string,
  locale: string,
  defaultCurrencyCode: string,
  wpVersion: string,
  gcpappkey: string,
  gcpapptoken: string,
  useXipay: boolean
  xipayUrl: string,
  xipayUsername: string,
  xipayPassword: string,
  xipayMerchantid: string,
  sapmid: string,
  sapgl: string,
  p24DelayToCancel: number,
  countryCodeMapping: [CountryCodeMapping]
}